FIL001-98 Restructure tests folder
