+++ tests/Fixtures/TestModels/TestModel.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Codedor\MediaLibrary\Tests\TestModels;
+namespace Codedor\MediaLibrary\Tests\Fixtures\TestModels;
 
 use Codedor\MediaLibrary\Interfaces\HasFormats;
 use Illuminate\Database\Eloquent\Model;
