+++ tests/Fixtures/TestModels/TestModelNotRegisterable.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Codedor\MediaLibrary\Tests\TestModels;
+namespace Codedor\MediaLibrary\Tests\Fixtures\TestModels;
 
 use Illuminate\Database\Eloquent\Model;
