+++ tests/Unit/Models/AttachmentTest.php
@@ -2,12 +2,32 @@
 
 use Codedor\MediaLibrary\Facades\Formats;
 use Codedor\MediaLibrary\Models\Attachment;
-use Codedor\MediaLibrary\Tests\TestFormats\TestHero;
+use Codedor\MediaLibrary\Tests\Fixtures\TestFormats\TestHero;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 
 uses(RefreshDatabase::class);
+
+it('returns the directory', function () {
+    /** @var Attachment $attachment */
+    $attachment = createAttachment([
+        'id' => 1,
+    ]);
+
+    expect($attachment->directory)
+        ->toBe('attachments/1');
+});
+
+it('returns the filename', function () {
+    /** @var Attachment $attachment */
+    $attachment = createAttachment([
+        'name' => 'test-file',
+    ]);
+
+    expect($attachment->filename)
+        ->toBe("test-file.$attachment->extension");
+});
 
 it('deletes root directory when attachment is removed from db', function () {
     Storage::fake('public');
@@ -21,7 +41,7 @@ it('deletes root directory when attachment is removed from db', function () {
 
     $attachment->getStorage()->put(
         $attachment->file_path,
-        File::get(__DIR__ . '/../../TestFiles/test.jpg')
+        File::get(__DIR__ . '/../../Fixtures/images/test.jpg')
     );
 
     $attachment->delete();
